+++ packages/radix-ui-themes/changelog.md
@@ -6,6 +6,8 @@
   - Tweak spacing
 - `Button`
   - Fix disabled ghost button padding
+- `ThemePanel`
+  - Improve tooltips content
 
 ## 0.0.43